LeadUp/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pydantic==2.10.2
pydantic-settings==2.6.1
PyJWT==2.9.0
bcrypt==4.2.0
slowapi==0.1.9
httpx==0.28.0
apscheduler==3.10.4
scrapling[fetchers]==0.2.9
python-dotenv==1.0.1
